Unifies local config format
