+++ SharedServices/SharedServices.csproj
@@ -10,4 +10,8 @@
       <ProjectReference Include="..\GatewayPluginContract\GatewayPluginContract.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <PackageReference Include="McMaster.NETCore.Plugins" Version="2.0.0" />
+    </ItemGroup>
+
 </Project>
